Move local storage functions out of missionSet class
