--- src/web/utils/local-storage.ts
@@ -1,9 +1,0 @@
-/**
- * Provides an array of all saved mission set names in localStorage, sorted alphabetically.
- *
- * @returns {string[]} Names of all saved missions sets
- */
-export function listSavedMissionSets() {
-    const allMissionSets = JSON.parse(localStorage.getItem("missionSets") || "{}");
-    return Object.keys(allMissionSets).sort((a, b) => a.localeCompare(b));
-}
